+++ datarepo-clienttests/src/main/resources/configs/ScalePods.json
@@ -1,7 +1,7 @@
 {
-  "name": "KubeConfig",
+  "name": "ScalePods",
   "description": "One user ingests many files into one dataset while we manipulate the kubernetes pods",
-  "serverSpecificationFile": "shdev.json",
+  "serverSpecificationFile": "localhost.json",
   "billingAccount": "00708C-45D19D-27AAFA",
   "kubernetes": {
     "numberOfInitialPods" : 3
@@ -9,11 +9,18 @@
   "application": {},
   "testScripts": [
     {
-      "name": "KillKubePods",
+      "name": "ScalePodsUpDown",
       "parameters": [29],
       "totalNumberToRun": 1,
       "numberToRunInParallel": 1,
-      "expectedTimeForEach": 100000000000000000,
+      "expectedTimeForEach": 300,
+      "expectedTimeForEachUnit": "SECONDS"
+    }, {
+      "name": "ScalePodsToZero",
+      "parameters": [29],
+      "totalNumberToRun": 1,
+      "numberToRunInParallel": 1,
+      "expectedTimeForEach": 300,
       "expectedTimeForEachUnit": "SECONDS"
     }
   ],
